graph, graphql: Move CacheStatus to graph crate
