+++ graph/src/data/query/mod.rs
@@ -1,7 +1,9 @@
+mod cache_status;
 mod error;
 mod query;
 mod result;
 
+pub use self::cache_status::CacheStatus;
 pub use self::error::{QueryError, QueryExecutionError};
 pub use self::query::{Query, QueryVariables};
 pub use self::result::QueryResult;
