+++ graphql/src/execution/execution.rs
@@ -9,11 +9,11 @@ use stable_hash::crypto::SetHasher;
 use stable_hash::prelude::*;
 use stable_hash::utils::stable_hash;
 use std::collections::{BTreeMap, HashMap, HashSet, VecDeque};
-use std::fmt;
 use std::iter;
 use std::sync::{Mutex, RwLock};
 use std::time::Instant;
 
+use graph::data::query::CacheStatus;
 use graph::prelude::*;
 use graph::util::lfu_cache::LfuCache;
 
@@ -273,39 +273,6 @@ fn cache_key(
         block_ptr,
     };
     stable_hash::<SetHasher, _>(&query)
-}
-
-/// Used for checking if a response hit the cache.
-#[derive(Copy, Clone)]
-pub(crate) enum CacheStatus {
-    /// Hit is a hit in the generational cache.
-    Hit,
-
-    /// Shared is a hit in the herd cache.
-    Shared,
-
-    /// Insert is a miss that inserted in the generational cache.
-    Insert,
-
-    /// A miss is none of the above.
-    Miss,
-}
-
-impl Default for CacheStatus {
-    fn default() -> Self {
-        CacheStatus::Miss
-    }
-}
-
-impl fmt::Display for CacheStatus {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            CacheStatus::Hit => f.write_str("hit"),
-            CacheStatus::Shared => f.write_str("shared"),
-            CacheStatus::Insert => f.write_str("insert"),
-            CacheStatus::Miss => f.write_str("miss"),
-        }
-    }
 }
 
 /// Contextual information passed around during query execution.
